requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
faiss-cpu
streamlit
openai
markdown